requirements.txt
pandas
python-dotenv
pinecone-client
openai
vercel-helpers